Move common API type alias to api/index
