+++ client/src/stores/datasetStore.ts
@@ -1,7 +1,8 @@
 import { defineStore } from "pinia";
 import Vue, { computed, ref } from "vue";
 
-import { DatasetDetails, DatasetEntry, HistoryContentItemBase } from "./services";
+import { DatasetDetails, DatasetEntry, HistoryContentItemBase } from "@/api";
+
 import { fetchDatasetDetails } from "./services/dataset.service";
 
 export const useDatasetStore = defineStore("datasetStore", () => {
